Portable URL implementation
This change intercepts the storage and retrieval of "content" and "desc"
fields.
It changes instnaces of the "webpath" and 'full webpath" into tokens on
saving an does the reverse on retrieving.

Note: if raw 'get' and 'put' methods are used for these fields anyware
those places will need to be updated to make these conversions.