cleanup use of base get/set methods
The base get/set methods should not be used when there are higher level
methods for the field. Many times this was to get the full database
field rather than just the current language string.

This change implements a $locale parameter of 'all' so that the higher
level methods can be used to get all the language strings.

The main reason for doing this at this time is the need to process these
gets and sets to handle the portable url tags.